// src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A date string, such as 2007-12-03 (YYYY-MM-DD), compliant with ISO 8601 standard for representation of dates using the Gregorian calendar. */
  Date: { input: unknown; output: unknown; }
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the date-timeformat outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representationof dates and times using the Gregorian calendar. */
  DateTime: { input: unknown; output: unknown; }
  Hex: { input: unknown; output: unknown; }
  /** Raw JSON value */
  Json: { input: unknown; output: unknown; }
  /** The Long scalar type represents non-fractional signed whole numeric values. Long can represent values between -(2^63) and 2^63 - 1. */
  Long: { input: unknown; output: unknown; }
  RGBAHue: { input: unknown; output: unknown; }
  RGBATransparency: { input: unknown; output: unknown; }
  /** Slate-compatible RichText AST */
  RichTextAST: { input: unknown; output: unknown; }
};

export type Aggregate = {
  count: Scalars['Int']['output'];
};

/** Asset system model */
export type Asset = Node & {
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  /** Get the document in other stages */
  documentInStages: Array<Asset>;
  /** The file name */
  fileName: Scalars['String']['output'];
  /** The file handle */
  handle: Scalars['String']['output'];
  /** The height of the file */
  height?: Maybe<Scalars['Float']['output']>;
  /** List of Asset versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  imageCollection: Array<Collection>;
  /** System Locale field */
  locale: Locale;
  /** Get the other localizations for this document */
  localizations: Array<Asset>;
  /** The mime type of the file */
  mimeType?: Maybe<Scalars['String']['output']>;
  productImages: Array<Product>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  /** The file size */
  size?: Maybe<Scalars['Float']['output']>;
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
  /** Get the url for the asset with provided transformations applied. */
  url: Scalars['String']['output'];
  /** The file width */
  width?: Maybe<Scalars['Float']['output']>;
};
// ...
export type Locale =
  /** System locale */
  | 'en';
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID']['output'];
  /** The Stage of an object */
  stage: Stage;
};
// ...
export type Product = Node & {
  averageRating?: Maybe<Scalars['Int']['output']>;
  categories: Array<Category>;
  collections: Array<Collection>;
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  description: Scalars['String']['output'];
  /** Get the document in other stages */
  documentInStages: Array<Product>;
  /** List of Product versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  images: Array<Asset>;
  /** System Locale field */
  locale: Locale;
  /** Get the other localizations for this document */
  localizations: Array<Product>;
  name: Scalars['String']['output'];
  orderItems: Array<OrderItem>;
  price: Scalars['Int']['output'];
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  reviews: Array<Review>;
  scheduledIn: Array<ScheduledOperation>;
  slug: Scalars['String']['output'];
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
  variants: Array<ProductVariants>;
};
// ...
export type ProductOrderByInput =
  | 'averageRating_ASC'
  | 'averageRating_DESC'
  | 'createdAt_ASC'
  | 'createdAt_DESC'
  | 'description_ASC'
  | 'description_DESC'
  | 'id_ASC'
  | 'id_DESC'
  | 'name_ASC'
  | 'name_DESC'
  | 'price_ASC'
  | 'price_DESC'
  | 'publishedAt_ASC'
  | 'publishedAt_DESC'
  | 'slug_ASC'
  | 'slug_DESC'
  | 'updatedAt_ASC'
  | 'updatedAt_DESC';
// ...
export type ProductVariants = ProductColorVariant | ProductSizeColorVariant | ProductSizeVariant;
// ...
export type Stage =
  /** The Draft is the default stage for all your content. */
  | 'DRAFT'
  /** The Published stage is where you can publish your content to. */
  | 'PUBLISHED';

export type SystemDateTimeFieldVariation =
  | 'BASE'
  | 'COMBINED'
  | 'LOCALIZATION';

export type UnpublishLocaleInput = {
  /** Locales to unpublish */
  locale: Locale;
  /** Stages to unpublish selected locales from */
  stages: Array<Stage>;
};

/** User system model */
export type User = Node & {
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** Get the document in other stages */
  documentInStages: Array<User>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  /** Flag to determine if user is active or not */
  isActive: Scalars['Boolean']['output'];
  /** User Kind. Can be either MEMBER, PAT or PUBLIC */
  kind: UserKind;
  /** The username */
  name: Scalars['String']['output'];
  /** Profile Picture url */
  picture?: Maybe<Scalars['String']['output']>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
};
// ...
export type Version = {
  createdAt: Scalars['DateTime']['output'];
  id: Scalars['ID']['output'];
  revision: Scalars['Int']['output'];
  stage: Stage;
};
// ...
export type ProductsCountQueryVariables = Exact<{ [key: string]: never; }>;


export type ProductsCountQuery = { productsConnection: { aggregate: { count: number } } };

export type ProductsCountByCategorySlugQueryVariables = Exact<{
  slug: Scalars['String']['input'];
}>;


export type ProductsCountByCategorySlugQuery = { categories: Array<{ id: string, name: string, slug: string, products: Array<{ id: string }> }> };

export type ProductGetListQueryVariables = Exact<{
  limit: Scalars['Int']['input'];
  offset: Scalars['Int']['input'];
  orderBy?: InputMaybe<ProductOrderByInput>;
}>;


export type ProductGetListQuery = { productsConnection: { products: Array<{ node: { id: string, name: string, price: number, averageRating?: number | null, categories: Array<{ name: string }>, images: Array<{ url: string, fileName: string }> } }>, pageInfo: { hasNextPage: boolean, hasPreviousPage: boolean } } };

export type ProductsGetListByCategorySlagQueryVariables = Exact<{
  slag: Scalars['String']['input'];
  limit: Scalars['Int']['input'];
  offset: Scalars['Int']['input'];
  orderBy?: InputMaybe<ProductOrderByInput>;
}>;


export type ProductsGetListByCategorySlagQuery = { productsConnection: { products: Array<{ node: { id: string, name: string, price: number, averageRating?: number | null, categories: Array<{ name: string }>, images: Array<{ url: string, fileName: string }> } }>, pageInfo: { hasNextPage: boolean, hasPreviousPage: boolean } } };

export type ProductsGetListByCollectionSlagQueryVariables = Exact<{
  slag: Scalars['String']['input'];
}>;


export type ProductsGetListByCollectionSlagQuery = { products: Array<{ id: string, name: string, price: number, averageRating?: number | null, categories: Array<{ name: string }>, images: Array<{ url: string, fileName: string }> }> };

export type ProductsGetListSearchQueryVariables = Exact<{
  search: Scalars['String']['input'];
}>;


export type ProductsGetListSearchQuery = { products: Array<{ id: string, name: string, price: number, averageRating?: number | null, categories: Array<{ name: string }>, images: Array<{ url: string, fileName: string }> }> };
// ...
export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];

  constructor(private value: string, public __meta__?: Record<string, any>) {
    super(value);
  }

  toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value;
  }
}
// ...
export const ProductsCountDocument = new TypedDocumentString(`
    query ProductsCount {
  productsConnection {
    aggregate {
      count
    }
  }
}
    `) as unknown as TypedDocumentString<ProductsCountQuery, ProductsCountQueryVariables>;
export const ProductsCountByCategorySlugDocument = new TypedDocumentString(`
    query ProductsCountByCategorySlug($slug: String!) {
  categories(where: {slug: $slug}) {
    id
    name
    slug
    products {
      id
    }
  }
}
    `) as unknown as TypedDocumentString<ProductsCountByCategorySlugQuery, ProductsCountByCategorySlugQueryVariables>;
export const ProductGetListDocument = new TypedDocumentString(`
    query ProductGetList($limit: Int!, $offset: Int!, $orderBy: ProductOrderByInput) {
  productsConnection(first: $limit, skip: $offset, orderBy: $orderBy) {
    products: edges {
      node {
        ...ProductListItem
      }
    }
    pageInfo {
      hasNextPage
      hasPreviousPage
    }
  }
}
    fragment ProductImage on Asset {
  url
  fileName
}
fragment ProductListItem on Product {
  id
  name
  price
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}`) as unknown as TypedDocumentString<ProductGetListQuery, ProductGetListQueryVariables>;
export const ProductsGetListByCategorySlagDocument = new TypedDocumentString(`
    query ProductsGetListByCategorySlag($slag: String!, $limit: Int!, $offset: Int!, $orderBy: ProductOrderByInput) {
  productsConnection(
    where: {categories_some: {slug: $slag}}
    first: $limit
    skip: $offset
    orderBy: $orderBy
  ) {
    products: edges {
      node {
        ...ProductListItem
      }
    }
    pageInfo {
      hasNextPage
      hasPreviousPage
    }
  }
}
    fragment ProductImage on Asset {
  url
  fileName
}
fragment ProductListItem on Product {
  id
  name
  price
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}`) as unknown as TypedDocumentString<ProductsGetListByCategorySlagQuery, ProductsGetListByCategorySlagQueryVariables>;
export const ProductsGetListByCollectionSlagDocument = new TypedDocumentString(`
    query ProductsGetListByCollectionSlag($slag: String!) {
  products(where: {collections_some: {slug: $slag}}) {
    ...ProductListItem
  }
}
    fragment ProductImage on Asset {
  url
  fileName
}
fragment ProductListItem on Product {
  id
  name
  price
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}`) as unknown as TypedDocumentString<ProductsGetListByCollectionSlagQuery, ProductsGetListByCollectionSlagQueryVariables>;
export const ProductsGetListSearchDocument = new TypedDocumentString(`
    query ProductsGetListSearch($search: String!) {
  products(where: {_search: $search}) {
    ...ProductListItem
  }
}
    fragment ProductImage on Asset {
  url
  fileName
}
fragment ProductListItem on Product {
  id
  name
  price
  categories(first: 1) {
    name
  }
  images(first: 1) {
    ...ProductImage
  }
  averageRating
}`) as unknown as TypedDocumentString<ProductsGetListSearchQuery, ProductsGetListSearchQueryVariables>;